@@ -4,78 +4,78 @@
   "license": "MIT",
   "scripts": {
     "ng": "nx",
-    "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2015 browser module main",
+    "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2020 browser module main",
     "start": "nx serve",
     "build": "nx build",
     "test": "nx test",
+    "upgrade-libs": "npm update @dontcode/core @dontcode/plugin-common @dontcode/sandbox",
+    "align-version": "PACKAGE_VERSION=$(npm view @dontcode/plugin-seed versions --json | jq -r '.[-1]') && cd libs/seed && npm version $PACKAGE_VERSION",
+    "local-publish-all": "cd libs/seed && npm version prerelease && cd ../.. && nx run seed:build:development && cd dist/libs/seed && npm publish --tag=next",
+    "local-publish": "cd libs/${npm_config_dir} && npm version prerelease && cd ../.. && nx run ${npm_config_dir}:build:development && cd dist/libs/${npm_config_dir} && npm publish --tag=next",
     "push-all": "cd libs/seed && npm version prerelease && cd ../.. && nx run seed:build:development && cd dist/libs/seed && yalc push",
     "push": "cd libs/${npm_config_dir} && npm version prerelease && cd ../.. && nx run ${npm_config_dir}:build:development && cd dist/libs/${npm_config_dir} && yalc push"
   },
   "private": true,
   "dependencies": {
-    "@angular-architects/module-federation-runtime": "~14.2.0",
-    "@angular/animations": "^13.3.2",
-    "@angular/cdk": "^13.3.2",
-    "@angular/common": "^13.3.2",
-    "@angular/compiler": "^13.3.2",
-    "@angular/core": "^13.3.2",
-    "@angular/forms": "^13.3.2",
-    "@angular/platform-browser": "^13.3.2",
-    "@angular/platform-browser-dynamic": "^13.3.2",
-    "@angular/router": "^13.3.2",
-    "@dontcode/core": "^0.5.41",
-    "@dontcode/plugin-common": "^0.5.41",
-    "@dontcode/sandbox": "^0.5.41",
-    "async-mutex": "^0.3.2",
-    "broadcast-channel": "^4.11.0",
-    "dexie": "^3.2.1",
-    "jsonpath-plus": "^7.0.0",
-    "primeflex": "^3.2.1",
-    "primeicons": "^5.0.0",
-    "primeng": "^13.4.1",
-    "rxjs": "^7.5.5",
-    "tslib": "^2.4.0",
-    "zone.js": "^0.11.5"
+    "@angular-architects/module-federation-runtime": "~15.0.3",
+    "@angular/animations": "^15.1.2",
+    "@angular/cdk": "^15.1.2",
+    "@angular/common": "^15.1.2",
+    "@angular/compiler": "^15.1.2",
+    "@angular/core": "^15.1.2",
+    "@angular/forms": "^15.1.2",
+    "@angular/platform-browser": "^15.1.2",
+    "@angular/platform-browser-dynamic": "^15.1.2",
+    "@angular/router": "^15.1.2",
+    "@dontcode/core": "^0.6.2",
+    "@dontcode/plugin-common": "^0.6.2",
+    "@dontcode/sandbox": "^0.6.2",
+    "primeflex": "^3.3.0",
+    "primeicons": "^6.0.1",
+    "primeng": "^15.1.1",
+    "rxjs": "~7.8.0",
+    "tslib": "^2.4.1",
+    "zone.js": "^0.12.0"
   },
   "devDependencies": {
-    "@angular-architects/module-federation": "~14.2.0",
-    "@angular-devkit/build-angular": "^13.3.2",
-    "@angular-eslint/eslint-plugin": "^13.2.1",
-    "@angular-eslint/eslint-plugin-template": "^13.2.1",
-    "@angular-eslint/template-parser": "^13.2.1",
-    "@angular/cli": "^13.3.2",
-    "@angular/compiler-cli": "^13.3.2",
-    "@angular/language-service": "^13.3.2",
-    "@nrwl/angular": "14.0.3",
-    "@nrwl/cli": "14.0.3",
-    "@nrwl/cypress": "14.0.3",
-    "@nrwl/eslint-plugin-nx": "14.0.3",
-    "@nrwl/jest": "14.0.3",
-    "@nrwl/linter": "14.0.3",
-    "@nrwl/tao": "14.0.3",
-    "@nrwl/workspace": "14.0.3",
-    "@types/jest": "^27.4.1",
-    "@types/node": "^16.11.9",
-    "@typescript-eslint/eslint-plugin": "^5.20.0",
-    "@typescript-eslint/parser": "^5.18.0",
+    "@angular-architects/module-federation": "~15.0.3",
+    "@angular-devkit/build-angular": "^15.1.3",
+    "@angular-eslint/eslint-plugin": "^15.2.0",
+    "@angular-eslint/eslint-plugin-template": "^15.2.0",
+    "@angular-eslint/template-parser": "^15.2.0",
+    "@angular/cli": "~15.1.0",
+    "@angular/compiler-cli": "^15.1.2",
+    "@angular/language-service": "^15.1.2",
+    "@nrwl/angular": "15.6.3",
+    "@nrwl/cli": "15.6.3",
+    "@nrwl/cypress": "15.6.3",
+    "@nrwl/eslint-plugin-nx": "15.6.3",
+    "@nrwl/jest": "15.6.3",
+    "@nrwl/linter": "15.6.3",
+    "@nrwl/tao": "15.6.3",
+    "@nrwl/workspace": "15.6.3",
+    "@types/jest": "^28.1.8",
+    "@types/node": "^18.11.18",
+    "@typescript-eslint/eslint-plugin": "^5.49.0",
+    "@typescript-eslint/parser": "^5.49.0",
     "codelyzer": "^6.0.2",
-    "cypress": "^9.5.4",
-    "dotenv": "^16.0.0",
-    "eslint": "^8.14.0",
-    "eslint-config-prettier": "^8.5.0",
+    "cypress": "^12.2.0",
+    "dotenv": "^16.0.3",
+    "eslint": "^8.33.0",
+    "eslint-config-prettier": "^8.6.0",
     "eslint-plugin-cypress": "^2.12.1",
-    "fake-indexeddb": "^3.1.7",
-    "jest": "^27.5.1",
-    "jest-preset-angular": "^11.1.1",
-    "ng-packagr": "^13.3.1",
-    "nx": "14.0.3",
-    "postcss": "^8.4.12",
-    "postcss-import": "^14.1.0",
-    "postcss-preset-env": "^7.4.3",
+    "jest": "^28.1.3",
+    "jest-environment-jsdom": "^28.1.3",
+    "jest-preset-angular": "^12.2.6",
+    "ng-packagr": "^15.1.1",
+    "nx": "15.6.3",
+    "postcss": "^8.4.21",
+    "postcss-import": "^15.1.0",
+    "postcss-preset-env": "^8.0.1",
     "postcss-url": "^10.1.3",
-    "prettier": "^2.6.2",
-    "ts-jest": "^27.1.4",
-    "ts-node": "^10.7.0",
-    "typescript": "^4.6.3"
+    "prettier": "^2.8.3",
+    "ts-jest": "^28.0.8",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.9.4"
   }
 }
